refactor(cli): rename util/workspace.ts to util/command.ts, inline lifecycle
Rename workspace.ts → command.ts to match what it actually contains
(CLI command infrastructure, not workspace logic) and avoid naming
collision with packages/workspace.

Inline openWorkspaceFromDisk and withWorkspace into runCommand—they
had no external callers. Delete OpenWorkspaceOptions and
OpenWorkspaceResult types (unused outside the file).

Public surface reduced from 7 exports to 4:
  defineCommand, withWorkspaceOptions, runCommand, resolveTable

Private resolveWorkspace helper extracts the find-by-id logic.
