+++ packages/cli/src/commands/kv.ts
@@ -8,7 +8,7 @@
 import type { Argv, CommandModule } from 'yargs';
 import { formatYargsOptions, outputError } from '../util/format-output';
 import { parseJsonInput, readStdinSync } from '../util/parse-input';
-import { defineCommand, runCommand, withWorkspaceOptions } from '../util/workspace';
+import { defineCommand, runCommand, withWorkspaceOptions } from '../util/command';
 
 /**
  * Parse a value from argv positional, --file, or stdin.
