+++ packages/cli/src/commands/run.ts
@@ -14,7 +14,7 @@ import {
 	defineCommand,
 	runCommand,
 	withWorkspaceOptions,
-} from '../util/workspace';
+} from '../util/command';
 
 /**
  * @example
